third_party/blink/web_tests/external/wpt/input-events/input-events-exec-command-expected.txt

This is a testharness.js-based test.
Found 70 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Calling execCommand("insertHorizontalRule", false, null) (inputType value)
  assert_equals: expected "insertHorizontalRule" but got ""
[FAIL] Calling execCommand("backColor", false, #FF0000) (inputType value)
  assert_equals: expected "formatBackColor" but got ""
[FAIL] Calling execCommand("backColor", false, #FF0000) (data value)
  assert_equals: expected (string) "rgb(255, 0, 0)" but got (object) null
[FAIL] Calling execCommand("backColor", false, #00FF00FF) (inputType value)
  assert_equals: expected "formatBackColor" but got ""
[FAIL] Calling execCommand("backColor", false, #00FF00FF) (data value)
  assert_equals: expected (string) "rgb(0, 255, 0)" but got (object) null
[FAIL] Calling execCommand("backColor", false, #0000FF88) (inputType value)
  assert_equals: expected "formatBackColor" but got ""
[FAIL] Calling execCommand("backColor", false, #0000FF88) (data value)
  assert_equals: expected (string) "rgba(0, 0, 255, 0.533)" but got (object) null
[FAIL] Calling execCommand("backColor", false, orange) (inputType value)
  assert_equals: expected "formatBackColor" but got ""
[FAIL] Calling execCommand("backColor", false, orange) (data value)
  assert_equals: expected (string) "rgb(255, 165, 0)" but got (object) null
[FAIL] Calling execCommand("backColor", false, Inherit) (inputType value)
  assert_equals: expected "formatBackColor" but got ""
[FAIL] Calling execCommand("backColor", false, Inherit) (data value)
  assert_equals: expected (string) "inherit" but got (object) null
[FAIL] Calling execCommand("backColor", false, Initial) (inputType value)
  assert_equals: expected "formatBackColor" but got ""
[FAIL] Calling execCommand("backColor", false, Initial) (data value)
  assert_equals: expected (string) "initial" but got (object) null
[FAIL] Calling execCommand("backColor", false, Reset) (inputType value)
  assert_equals: expected "formatBackColor" but got ""
[FAIL] Calling execCommand("backColor", false, Reset) (data value)
  assert_equals: expected (string) "reset" but got (object) null
[FAIL] Calling execCommand("backColor", false, transparent) (inputType value)
  assert_equals: expected "formatBackColor" but got ""
[FAIL] Calling execCommand("backColor", false, transparent) (data value)
  assert_equals: expected (string) "rgba(0, 0, 0, 0)" but got (object) null
[FAIL] Calling execCommand("backColor", false, CurrentColor) (inputType value)
  assert_equals: expected "formatBackColor" but got ""
[FAIL] Calling execCommand("backColor", false, CurrentColor) (data value)
  assert_equals: expected (string) "currentcolor" but got (object) null
[FAIL] Calling execCommand("backColor", false, Invalid-Value) (inputType value)
  assert_equals: expected "formatBackColor" but got ""
[FAIL] Calling execCommand("backColor", false, Invalid-Value) (data value)
  assert_equals: expected (string) "Invalid-Value" but got (object) null
[FAIL] Calling execCommand("foreColor", false, #FF0000) (inputType value)
  assert_equals: expected "formatFontColor" but got ""
[FAIL] Calling execCommand("foreColor", false, #FF0000) (data value)
  assert_equals: expected (string) "rgb(255, 0, 0)" but got (object) null
[FAIL] Calling execCommand("foreColor", false, #00FF00FF) (inputType value)
  assert_equals: expected "formatFontColor" but got ""
[FAIL] Calling execCommand("foreColor", false, #00FF00FF) (data value)
  assert_equals: expected (string) "rgb(0, 255, 0)" but got (object) null
[FAIL] Calling execCommand("foreColor", false, #0000FF88) (inputType value)
  assert_equals: expected "formatFontColor" but got ""
[FAIL] Calling execCommand("foreColor", false, #0000FF88) (data value)
  assert_equals: expected (string) "rgba(0, 0, 255, 0.533)" but got (object) null
[FAIL] Calling execCommand("foreColor", false, orange) (inputType value)
  assert_equals: expected "formatFontColor" but got ""
[FAIL] Calling execCommand("foreColor", false, orange) (data value)
  assert_equals: expected (string) "rgb(255, 165, 0)" but got (object) null
[FAIL] Calling execCommand("foreColor", false, Inherit) (inputType value)
  assert_equals: expected "formatFontColor" but got ""
[FAIL] Calling execCommand("foreColor", false, Inherit) (data value)
  assert_equals: expected (string) "inherit" but got (object) null
[FAIL] Calling execCommand("foreColor", false, Initial) (inputType value)
  assert_equals: expected "formatFontColor" but got ""
[FAIL] Calling execCommand("foreColor", false, Initial) (data value)
  assert_equals: expected (string) "initial" but got (object) null
[FAIL] Calling execCommand("foreColor", false, Reset) (inputType value)
  assert_equals: expected "formatFontColor" but got ""
[FAIL] Calling execCommand("foreColor", false, Reset) (data value)
  assert_equals: expected (string) "reset" but got (object) null
[FAIL] Calling execCommand("foreColor", false, transparent) (inputType value)
  assert_equals: expected "formatFontColor" but got ""
[FAIL] Calling execCommand("foreColor", false, transparent) (data value)
  assert_equals: expected (string) "rgba(0, 0, 0, 0)" but got (object) null
[FAIL] Calling execCommand("foreColor", false, CurrentColor) (inputType value)
  assert_equals: expected "formatFontColor" but got ""
[FAIL] Calling execCommand("foreColor", false, CurrentColor) (data value)
  assert_equals: expected (string) "currentcolor" but got (object) null
[FAIL] Calling execCommand("foreColor", false, Invalid-Value) (inputType value)
  assert_equals: expected "formatFontColor" but got ""
[FAIL] Calling execCommand("foreColor", false, Invalid-Value) (data value)
  assert_equals: expected (string) "Invalid-Value" but got (object) null
[FAIL] Calling execCommand("hiliteColor", false, #FF0000) (inputType value)
  assert_equals: expected "formatBackColor" but got ""
[FAIL] Calling execCommand("hiliteColor", false, #FF0000) (data value)
  assert_equals: expected (string) "rgb(255, 0, 0)" but got (object) null
[FAIL] Calling execCommand("hiliteColor", false, #00FF00FF) (inputType value)
  assert_equals: expected "formatBackColor" but got ""
[FAIL] Calling execCommand("hiliteColor", false, #00FF00FF) (data value)
  assert_equals: expected (string) "rgb(0, 255, 0)" but got (object) null
[FAIL] Calling execCommand("hiliteColor", false, #0000FF88) (inputType value)
  assert_equals: expected "formatBackColor" but got ""
[FAIL] Calling execCommand("hiliteColor", false, #0000FF88) (data value)
  assert_equals: expected (string) "rgba(0, 0, 255, 0.533)" but got (object) null
[FAIL] Calling execCommand("hiliteColor", false, orange) (inputType value)
  assert_equals: expected "formatBackColor" but got ""
[FAIL] Calling execCommand("hiliteColor", false, orange) (data value)
  assert_equals: expected (string) "rgb(255, 165, 0)" but got (object) null
[FAIL] Calling execCommand("hiliteColor", false, Inherit) (inputType value)
  assert_equals: expected "formatBackColor" but got ""
[FAIL] Calling execCommand("hiliteColor", false, Inherit) (data value)
  assert_equals: expected (string) "inherit" but got (object) null
[FAIL] Calling execCommand("hiliteColor", false, Initial) (inputType value)
  assert_equals: expected "formatBackColor" but got ""
[FAIL] Calling execCommand("hiliteColor", false, Initial) (data value)
  assert_equals: expected (string) "initial" but got (object) null
[FAIL] Calling execCommand("hiliteColor", false, Reset) (inputType value)
  assert_equals: expected "formatBackColor" but got ""
[FAIL] Calling execCommand("hiliteColor", false, Reset) (data value)
  assert_equals: expected (string) "reset" but got (object) null
[FAIL] Calling execCommand("hiliteColor", false, transparent) (inputType value)
  assert_equals: expected "formatBackColor" but got ""
[FAIL] Calling execCommand("hiliteColor", false, transparent) (data value)
  assert_equals: expected (string) "rgba(0, 0, 0, 0)" but got (object) null
[FAIL] Calling execCommand("hiliteColor", false, CurrentColor) (inputType value)
  assert_equals: expected "formatBackColor" but got ""
[FAIL] Calling execCommand("hiliteColor", false, CurrentColor) (data value)
  assert_equals: expected (string) "currentcolor" but got (object) null
[FAIL] Calling execCommand("hiliteColor", false, Invalid-Value) (inputType value)
  assert_equals: expected "formatBackColor" but got ""
[FAIL] Calling execCommand("hiliteColor", false, Invalid-Value) (data value)
  assert_equals: expected (string) "Invalid-Value" but got (object) null
[FAIL] Calling execCommand("fontName", false, monospace) (inputType value)
  assert_equals: expected "formatFontName" but got ""
[FAIL] Calling execCommand("fontName", false, monospace) (data value)
  assert_equals: expected (string) "monospace" but got (object) null
[FAIL] Calling execCommand("fontName", false,  monospace ) (inputType value)
  assert_equals: expected "formatFontName" but got ""
[FAIL] Calling execCommand("fontName", false,  monospace ) (data value)
  assert_equals: expected (string) " monospace " but got (object) null
[FAIL] Calling execCommand("fontName", false,   monospace  ) (inputType value)
  assert_equals: expected "formatFontName" but got ""
[FAIL] Calling execCommand("fontName", false,   monospace  ) (data value)
  assert_equals: expected (string) "  monospace  " but got (object) null
[FAIL] Calling execCommand("paste", false, null) (dataTransfer value, text/plain)
  assert_true: calling dataTransfer.getData(text/plain) caused exception: TypeError: Cannot read properties of null (reading 'getData') expected true got false
[FAIL] Calling execCommand("paste", false, null) (dataTransfer.clearData(text/plain))
  assert_true: calling dataTransfer.clearData(text/plain) caused exception: TypeError: Cannot read properties of null (reading 'clearData') expected true got false
[FAIL] Calling execCommand("paste", false, null) (dataTransfer.setData(text/plain))
  assert_true: calling dataTransfer.setData(text/plain) caused exception: TypeError: Cannot read properties of null (reading 'setData') expected true got false
Harness: the test ran to completion.
